Kernel: Move Thread's definition to the header file
